// src/shared/services/api/models/postApiV1AssessmentBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export type PostApiV1AssessmentBody = {
  CourseId?: string;
  Title?: string;
  Description?: string;
  PassingMark?: number;
  DurationMinute?: number;
  QuestionFile?: Blob;
};