// src/shared/services/api/models/updateAssessmentRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface UpdateAssessmentRequestDto {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  total_mark?: number | null;
  /** @nullable */
  passing_mark?: number | null;
  /** @nullable */
  duration_minute?: number | null;
}